update : fix process uptime and memory alignments
In processes list, the uptime is reversely sorted by string sort, which may be incorrectly sorted, for example

  9:01:23
  3 day 1:12:23
  21:02:45

Which is correct if by string sort, but not what we want. This patch adds a format function to pad the strings, so it can be sorted correct by string sort. However, a proc_info['rawtime'] is added for time delta as sorting key even it's not necessary with this new format function, it's safer to sort the raw data than its representation. The drawback is the uptime is fixed in English.

The alignment for memory usage percentage is also fixed with proper padding and float type.